Added `Dockerfile`, switched CI to use `docker build` for tests
The idea is that the same `Dockerfile` used for CI can be used for
running the workshop, so we roll with that.

That means a less nice CI pipeline, but good enough for our purposes.

Note how CI does **NOT** push the docker image: since this is a public
repository, we cannot really perform docker registry pushes consistently,
since the pipeline would always fail due to missing secrets.

Also, we had to revert the SQLite JSON access syntax: not yet available
for PHP + Ubuntu 22.04.